bug: fix symlinks and tests on Windows
Testing on Windows now works on Git Bash (so most Unix commands can still be used). Main problem in tests was that long paths can be created but not removed - rimraf somehow manages to do this. Functionally the main improvement is that we create a junction on Windows, so that we don't have to use the copy strategy, which is a lot slower.